move all controllers for reset password to passwordReset.js
